+++ src/main/java/roomescape/config/LoginMemberArgumentResolver.java
@@ -6,10 +6,10 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.member.domain.Member;
-import roomescape.member.service.MemberService;
 import roomescape.jwt.CookieTokenExtractor;
 import roomescape.jwt.TokenProvider;
+import roomescape.member.domain.Member;
+import roomescape.member.service.MemberService;
 
 public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolver {
