+++ src/main/java/roomescape/config/WebMvcConfig.java
@@ -5,8 +5,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.member.service.MemberService;
 import roomescape.jwt.TokenProvider;
+import roomescape.member.service.MemberService;
 
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
