+++ src/main/java/roomescape/view/controller/UserPageController.java
@@ -28,7 +28,6 @@ public class UserPageController {
 
     @GetMapping("/reservation-mine")
     public String reservationMinePage() {
-        //TODO - 로그인 못하면 /login 페이지로 redirect
         return "reservation-mine";
     }
 }
