--- src/main/java/roomescape/waiting/domain/WaitingWithRank.java
@@ -1,20 +1,0 @@
-package roomescape.waiting.domain;
-
-public class WaitingWithRank {
-
-    private final Waiting waiting;
-    private final Long rank;
-
-    public WaitingWithRank(Waiting waiting, Long rank) {
-        this.waiting = waiting;
-        this.rank = rank + 1;
-    }
-
-    public Waiting getWaiting() {
-        return waiting;
-    }
-
-    public Long getRank() {
-        return rank;
-    }
-}
