+++ src/main/java/roomescape/waiting/dto/WaitingWithRank.java
@@ -1,0 +1,27 @@
+package roomescape.waiting.dto;
+
+import roomescape.waiting.domain.Waiting;
+import roomescape.waiting.domain.WaitingStatus;
+
+public class WaitingWithRank {
+
+    private final Waiting waiting;
+    private final Long rank;
+
+    public WaitingWithRank(Waiting waiting, Long rank) {
+        this.waiting = waiting;
+        this.rank = rank + 1;
+    }
+
+    public Waiting getWaiting() {
+        return waiting;
+    }
+
+    public String getLabel() {
+        WaitingStatus status = waiting.getStatus();
+        if(status == WaitingStatus.PENDING) {
+            return rank + "번째 " + status.getLabel();
+        }
+        return status.getLabel();
+    }
+}
